Added the float type converter.

+++ ProgramDesign/project2_travel.c
@@ -7,10 +7,9 @@
 
 #include <stdio.h>
 #include <math.h>
-#include <tgmath.h>
 
 // A macro for the log(3/2):
-#define DENO log((float) 3 / 2)  
+#define DENO logf((float) 3 / 2)  
 
 // Functions:
 void days(void);
@@ -39,7 +38,7 @@ void days(void)
 		scanf("%hd", &second);
 		
 		// Calculate the days:
-		days = (short) trunc((log(first/second) / DENO) + 1);
+		days = (short) truncf((logf(first/second) / DENO) + 1);
 	
 		// Output the result:
 		printf("Case #%hd: %hd day(s)\n", cases, days);
